+++ src/Perimeterx.php
@@ -25,6 +25,7 @@
 
 namespace Perimeterx;
 
+
 use Psr\Log\LoggerInterface;
 
 final class Perimeterx
@@ -93,7 +94,7 @@ final class Perimeterx
                 'max_buffer_len' => 1,
                 'send_page_activities' => true,
                 'send_block_activities' => true,
-                'sdk_name' => 'PHP SDK v3.6.0',
+                'sdk_name' => 'PHP SDK v3.7.0',
                 'debug_mode' => false,
                 'perimeterx_server_host' => 'https://sapi-' . strtolower($pxConfig['app_id']) . '.perimeterx.net',
                 'captcha_script_host' => 'https://captcha.px-cdn.net',
@@ -107,7 +108,9 @@ final class Perimeterx
                 'ip_headers' => [],
                 'bypass_monitor_header' => null,
                 'custom_block_url' => null,
-                'defer_activities' => true
+                'defer_activities' => true,
+                'enable_json_response' => false,
+                'return_response' => false
             ], $pxConfig);
 
             if (empty($this->pxConfig['logger'])) {
@@ -127,6 +130,7 @@ final class Perimeterx
     public function pxVerify()
     {
         $pxCtx = null;
+        $extractedCredentials = null;
         $this->pxConfig['logger']->debug('Starting request verification');
         try {
             if (!$this->pxConfig['module_enabled']) {
@@ -191,7 +195,7 @@ final class Perimeterx
 
     /**
      * @param PerimeterxContext $pxCtx
-     * @return bool - a true value when user is scored ok/blocking is disabled
+     * @return mixed object|boolean as the verification result
      */
     private function handleVerification($pxCtx)
     {
@@ -216,10 +220,14 @@ final class Perimeterx
             return 1;
         }
 
-        $should_bypass_monitor = isset($this->pxConfig['bypass_monitor_header']) && isset($pxCtx->getHeaders()[$this->pxConfig['bypass_monitor_header']]) && $pxCtx->getHeaders()[$this->pxConfig['bypass_monitor_header']] == "1";
+        $headers = array_change_key_case($pxCtx->getHeaders(), CASE_LOWER);
+        $should_bypass_monitor = isset($this->pxConfig['bypass_monitor_header']) && isset($headers[strtolower($this->pxConfig['bypass_monitor_header'])]) && $headers[strtolower($this->pxConfig['bypass_monitor_header'])] == "1";
         if ($this->pxConfig['module_mode'] != Perimeterx::$ACTIVE_MODE && !$should_bypass_monitor ) {
             return 1;
         }
+
+        $accept_header = isset($_SERVER['HTTP_ACCEPT']) ? $_SERVER['HTTP_ACCEPT'] : $_SERVER['HTTP_CONTENT_TYPE'];
+        $is_json_response = $pxCtx->getCookieOrigin() == 'cookie' && strpos($accept_header,'application/json') !== false && $this->pxConfig['enable_json_response'];
 
         $block_uuid = $pxCtx->getUuid();
         $mustache = new \Mustache_Engine(array(
@@ -260,14 +268,38 @@ final class Perimeterx
                 header('Location: '.$page_url, true, 307);
                 die();
             } else {
-                $html = $mustache->render('block_template', $templateInputs);
-                $this->pxConfig['logger']->debug("Enforcing action: {$pxCtx->getBlockAction()} page is served");
+                if ($is_json_response == false) {
+                    $html = $mustache->render('block_template', $templateInputs);
+                    $this->pxConfig['logger']->debug("Enforcing action: {$pxCtx->getBlockAction()} page is served");
+                } else {
+                    $this->pxConfig['logger']->debug("Enforcing action: advanced blocking response is served");
+                }
             }
         }
 
         if ($pxCtx->getCookieOrigin() == 'cookie') {
-            header("Content-Type: text/html");
-            echo $html;
+            if($is_json_response) {
+                header("Content-Type: application/json");
+                $result = array(
+                    'appId' => $this->pxConfig['app_id'],
+                    'jsClientSrc' => $templateInputs['jsClientSrc'],
+                    'firstPartyEnabled' => false,
+                    'vid' => $templateInputs['vid'],
+                    'uuid' => $templateInputs['uuid'],
+                    'hostUr' => $templateInputs['hostUrl'],
+                    'blockScript' => $templateInputs['blockScript']
+                );
+                if ($this->pxConfig['return_response']) {
+                    return $result;
+                }
+                echo json_encode($result);
+            } else {
+                header("Content-Type: text/html");
+                if ($this->pxConfig['return_response']) {
+                    return $html;
+                }
+                echo $html;
+            }
         } else {
             header("Content-Type: application/json");
             $result = array(
@@ -278,6 +310,9 @@ final class Perimeterx
                 'page' => base64_encode($html),
                 'collectorUrl' => $collectorUrl
             );
+            if ($this->pxConfig['return_response']) {
+                return $result;
+            }
             echo json_encode($result);
         }
         die();