+++ src/PerimeterxActivitiesClient.php
@@ -76,7 +76,9 @@ class PerimeterxActivitiesClient
         $pxData['px_app_id'] = $this->pxConfig['app_id'];
         $pxData['url'] = $pxCtx->getFullUrl();
         $pxData['details'] = $details;
-        $this->addAdditionalFieldsToDetails($pxData['details'], $pxCtx->getAdditionalFields());
+
+        $additionalFields = $pxCtx->getAdditionalFields();
+        $this->addAdditionalFieldsToDetails($pxData['details'], $additionalFields);
 
         $vid = $pxCtx->getVid();
         if (isset($vid)) {
@@ -158,6 +160,12 @@ class PerimeterxActivitiesClient
     }
 
     private function addAdditionalFieldsToDetails(&$details, &$additionalFields) {
+        if (!function_exists('is_iterable')) {
+            function is_iterable($var)
+            {
+                return is_array($var) || $var instanceof \Traversable;
+            }
+        }
         if (!is_iterable($additionalFields)) {
             return;
         }